Remove loop in TimeManagement::init() and tweak time management logic
Allocate increment in proportion to the move importance of future moves.
The loop over hypMTG should be completely unnecessary with the new logic.
Instead of DBL_MIN, the asymptotic behavior of move_importance is exploited
to avoid issues with loss of precision, or underflow in extreme cases.
Move Overhead and Minimum Thinking Time are treated in a different way.